SILVER-TIER/ai_employee/pyproject.toml
[project]
name = "ai-employee-silver"
version = "2.0.0"
requires-python = ">=3.13"
dependencies = [
    "watchdog>=4.0.0",
    "python-dotenv>=1.0.0",
    "google-auth>=2.0.0",
    "google-auth-oauthlib>=1.0.0",
    "google-api-python-client>=2.0.0",
    "schedule>=1.2.0",

    # Silver Tier — NEW additions:
    "linkedin-api>=2.0.0",
    "requests>=2.31.0",
    "python-crontab>=3.0.0",
    "apscheduler>=3.10.0",
    "httpx>=0.27.0",
    "beautifulsoup4>=4.12.0",
]

[tool.hatch.build.targets.wheel]
packages = ["."]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
